backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
pymysql==1.1.0
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
email-validator==2.1.1
langchain==0.1.13
openai==1.14.2
scikit-learn==1.4.1.post1
numpy==1.26.4
pandas==2.2.1
yfinance==0.2.37
cachetools==5.3.3

